+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [0.16.0] - 2022-05-05
 ### Added
 - Build suggestions from unmatched gherkin steps ([#40](https://github.com/cucumber/language-service/pull/40))
 
@@ -133,7 +134,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ([#1732](https://github.com/cucumber/common/pull/1732)
 [aslakhellesoy](https://github.com/aslakhellesoy))
 
-[Unreleased]: https://github.com/cucumber/language-service/compare/v0.15.0...HEAD
+[Unreleased]: https://github.com/cucumber/language-service/compare/v0.16.0...HEAD
+[0.16.0]: https://github.com/cucumber/language-service/compare/v0.15.0...v0.16.0
 [0.15.0]: https://github.com/cucumber/language-service/compare/v0.14.4...v0.15.0
 [0.14.4]: https://github.com/cucumber/language-service/compare/v0.14.3...v0.14.4
 [0.14.3]: https://github.com/cucumber/language-service/compare/v0.14.2...v0.14.3